+++ src/ChoyUtilities/Assets/ChoyUtilities/Common/CHANGELOG.md
@@ -24,6 +24,7 @@ Changed
 Removed
 * Microphone Recorder, moved to Misc
 * Microphone Attribute, moved to Misc
+* Quaternion cast from Floater
 
 ## [1.0.2] - 2026-04-16